travel-planner-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.98.0
python-multipart==0.0.6
